+++ locals-tags.tf
@@ -3,4 +3,5 @@ locals {
     env   = var.environment
     stack = var.stack
   } : {}
+  tags = merge(local.default_tags, var.extra_tags)
 }