+++ r-naming.tf
@@ -1,10 +1,8 @@
 resource "azurecaf_name" "eventhub_namespace" {
-  for_each = var.eventhub_namespaces_hubs
-
   name          = var.stack
   resource_type = "azurerm_eventhub_namespace"
   prefixes      = var.name_prefix == "" ? null : [local.name_prefix]
-  suffixes      = compact([var.client_name, var.location_short, var.environment, each.key, local.name_suffix, var.use_caf_naming ? "" : "eh"])
+  suffixes      = compact([var.client_name, var.location_short, var.environment, local.name_suffix, var.use_caf_naming ? "" : "eh"])
   use_slug      = var.use_caf_naming
   clean_input   = true
   separator     = "-"