+++ variables.tf
@@ -1,70 +1,84 @@
-variable "location" {
-  description = "Azure region"
+# Generic variables
+
+variable "client_name" {
+  description = "Client name/account used in naming."
   type        = string
-  default     = ""
 }
 
 variable "environment" {
-  description = "Project environment"
+  description = "Project environment."
   type        = string
 }
 
-variable "extra_tags" {
-  description = "Extra tags"
-  type        = map(string)
-  default     = {}
+variable "stack" {
+  description = "Project stack name."
+  type        = string
 }
 
 variable "resource_group_name" {
-  description = "Azure Resource Group name"
+  description = "Name of the resource group."
   type        = string
 }
 
-# EVT-HUB-01
-variable "namespace_is_public" {
-  description = "Specify the current use case (sg_Resource_ControlTower_Profile=Public or Private)."
-  type        = bool
-  default     = false
+variable "location" {
+  description = "Azure location for Eventhub."
+  type        = string
 }
 
-variable "namespace_sku" {
-  description = "Defines which tier to use. Valid options are Basic, Standard, and Premium. Please not that setting this field to Premium will force the creation of a new resource and also requires setting zone_redundant to true."
+variable "location_short" {
+  description = "Short string for Azure location."
   type        = string
+}
+
+# EventHub Namespace
+variable "namespace_parameters" {
+  description = <<EOD
+EventHub namespace parameters.
+ * sku:                  Defines which tier to use. Valid options are Basic, Standard, and Premium. Please not that setting this field to Premium will force the creation of a new resource and also requires setting zone_redundant to true.
+ * capacity:             Specifies the Capacity / Throughput Units for a Standard SKU namespace. Default capacity has a maximum of 2, but can be increased in blocks of 2 on a committed purchase basis.
+ * auto_inflate_enabled: Is Auto Inflate enabled for the Event Hub namespace?
+ * dedicated_cluster_id: Specifies the ID of the Event Hub Dedicated Cluster where this namespace should created.
+ * maximum_throughput_units: Specifies the maximum number of throughput units when Auto Inflate is Enabled. Valid values range from 1 - 20.
+ * zone_redundant:       Specifies if the Event Hub namespace should be Zone Redundant (created across Availability Zones). Changing this forces a new resource to be created.
+ * local_authentication_enabled: Is SAS authentication enabled for the EventHub Namespace?
+ * public_network_access_enabled: Is public network access enabled for the EventHub Namespace? Defaults to `true`.
+ * minimum_tls_version:  The minimum supported TLS version for this EventHub Namespace. Valid values are: `1.0`, `1.1` and `1.2`. The current default minimum TLS version is `1.2`.
+EOD
+  type = object({
+    sku                           = string
+    capacity                      = optional(number, 2)
+    auto_inflate_enabled          = optional(bool, false)
+    dedicated_cluster_id          = optional(string)
+    maximum_throughput_units      = optional(number)
+    zone_redundant                = optional(bool, true)
+    local_authentication_enabled  = optional(bool)
+    public_network_access_enabled = optional(bool, true)
+    minimum_tls_version           = optional(string, "1.2")
+  })
   validation {
-    condition     = contains(["Basic", "Standard", "Premium"], var.namespace_sku)
+    condition     = contains(["Basic", "Standard", "Premium"], var.namespace_parameters.sku)
     error_message = "Sku value is invalid. Must be `Basic`, `Standard` or `Premium`."
   }
 }
 
-variable "namespace_capacity" {
-  description = "Specifies the Capacity / Throughput Units for a Standard SKU namespace. Default capacity has a maximum of 2, but can be increased in blocks of 2 on a committed purchase basis."
-  type        = number
-  default     = 2
+variable "namespace_network_rules" {
+  description = "`network_rulesets` map block as defined below."
+  default     = {}
+  type = map(object({
+    default_action                 = optional(string, "Deny")
+    trusted_service_access_enabled = optional(bool, true)
+    virtual_network_rules = optional(list(object({
+      subnet_id                                       = string
+      ignore_missing_virtual_network_service_endpoint = optional(bool, false)
+    })), [])
+    ip_rules = optional(list(object({
+      ip_mask = string
+      action  = optional(string, "Allow")
+    })), [])
+  }))
 }
 
-variable "namespace_auto_inflate_enabled" {
-  description = "Is Auto Inflate enabled for the Event Hub namespace ?"
-  type        = bool
-  default     = false
-}
-
-variable "namespace_dedicated_cluster_id" {
-  description = "Specifies the ID of the Event Hub Dedicated Cluster where this namespace should created."
-  type        = string
-  default     = null
-}
-
-variable "namespace_maximum_throughput_units" {
-  description = "Specifies the maximum number of throughput units when Auto Inflate is Enabled. Valid values range from 1 - 20."
-  type        = number
-  default     = null
-}
-
-variable "namespace_zone_redundant" {
-  description = "Specifies if the Event Hub namespace should be Zone Redundant (created across Availability Zones). Changing this forces a new resource to be created."
-  type        = bool
-  default     = true
-}
+# EventHubs
 
 variable "partition_count" {
   description = "Specifies the current number of shards on the Event Hub."